# frontend/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KYC API",
    "version": "1.0.0",
    "description": "JSON:API endpoints for Applicant accounts and sessions, KYC applications and document evidence, reviewer workload, and access checks. Authenticated endpoints use the first-party session cookie. Error responses use a JSON:API errors array with safe, client-actionable details."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current frontend origin; /api/v1 is proxied to the Java API in development."
    }
  ],
...
  "paths": {
...
    "/api/v1/applicant-applications": {
      "post": {
        "operationId": "createApplicantApplication",
        "responses": {
          "200": {
            "description": "The Applicant?s existing active draft, returned without creating a duplicate.",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicantApplicationDocument"
                }
              }
            }
          },
          "201": {
            "description": "A new draft was created.",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string",
                  "format": "uri-reference"
                },
                "description": "URI of the newly created application resource."
              }
            },
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicantApplicationDocument"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/AccessDenied"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": ["Applicant applications"],
        "summary": "Create or retrieve the current Applicant draft",
        "description": "Starts a draft when none exists. Returns the existing active draft when one already exists, so repeated calls do not create duplicate current drafts. Send the session cookie and, when present, its CSRF token.",
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
    "/api/v1/applicant-applications/{applicationId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ApplicationId"
        }
      ],
      "get": {
        "operationId": "getApplicantApplicationStatus",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "The authenticated Applicant's minimized application lifecycle state.",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationDocument"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "404": {
            "$ref": "#/components/responses/ApplicationNotFound"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": ["Applicant applications"],
        "summary": "Get Applicant application status",
        "description": "Returns the authenticated Applicant?s minimized application lifecycle state. The resource omits personal form answers and document contents."
      },
      "patch": {
        "operationId": "submitApplicantApplication",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/vnd.api+json": {
              "schema": {
                "$ref": "#/components/schemas/SubmissionRequestDocument"
              }
            }
          },
          "description": "JSON:API document confirming submission and carrying the expected form version returned by the latest form read."
        },
        "responses": {
          "200": {
            "description": "Submission committed or the same already-committed submission returned after a replay.",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationDocument"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/MalformedJsonApiDocument"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/AccessDenied"
          },
          "404": {
            "$ref": "#/components/responses/ApplicationNotFound"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "409": {
            "$ref": "#/components/responses/FormVersionConflict"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedRequestMediaType"
          },
          "422": {
            "$ref": "#/components/responses/SubmissionUnprocessable"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": ["Applicant applications"],
        "parameters": [
          {
            "name": "X-XSRF-TOKEN",
            "in": "header",
            "required": false,
            "description": "CSRF token from the XSRF-TOKEN cookie. Required when sending a session cookie with a state-changing request.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "summary": "Submit the Applicant application",
        "description": "Submits a complete draft after explicit confirmation and an expected form version. A successful replay returns the already-committed submission. Incomplete applications return 422 with missing field names in the error detail; stale versions or invalid lifecycle state return 409."
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-KYCSESSION",
        "description": "Opaque, secure, HttpOnly cookie set after successful sign-in. The browser sends it automatically to the API origin."
      }
    },
    "schemas": {
...
      "ApplicantApplicationDocument": {
        "type": "object",
        "required": ["data"],
        "properties": {
          "data": {
            "type": "object",
            "required": ["type", "id", "attributes"],
            "properties": {
              "type": {
                "const": "applicant-applications"
              },
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "attributes": {
                "type": "object",
                "required": ["status", "currentStep", "progress", "href"],
                "properties": {
                  "status": {
                    "const": "draft"
                  },
                  "currentStep": {
                    "enum": ["personal-details", "identity-and-address"],
                    "description": "Step currently open in the Applicant journey."
                  },
                  "href": {
                    "type": "string",
                    "format": "uri-reference"
                  },
                  "progress": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": ["step", "state"],
                      "properties": {
                        "step": {
                          "enum": ["personal-details", "identity-and-address"]
                        },
                        "state": {
                          "enum": ["not-started", "current", "complete"]
                        }
                      }
                    },
                    "description": "Completion state for each application step."
                  }
                }
              }
            }
          }
        },
        "description": "JSON:API response describing the Applicant?s current draft and step progress."
      },
      "ErrorDocument": {
        "type": "object",
        "required": ["errors"],
        "properties": {
          "errors": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "description": "HTTP status code represented as a string, such as \"422\"."
                },
                "code": {
                  "type": "string",
                  "description": "Stable machine-readable error code for client handling."
                },
                "title": {
                  "type": "string",
                  "description": "Short, human-readable summary of the error."
                },
                "detail": {
                  "type": "string",
                  "description": "Safe, actionable explanation. It does not contain credentials, internal exception details, or stored document contents."
                },
                "source": {
                  "type": "object",
                  "properties": {
                    "pointer": {
                      "type": "string",
                      "description": "JSON Pointer to the invalid member in the request document, such as /data/attributes/email."
                    }
                  },
                  "description": "Location of the invalid request member. Omitted when the error is not tied to one field.",
                  "required": ["pointer"]
                }
              },
              "required": ["status", "code", "title", "detail"],
              "description": "Safe JSON:API error details. source is present for field validation errors and points to the invalid request member."
            },
            "description": "One or more error objects. Each object describes one failure; validation can return multiple errors in the same response."
          }
        },
        "description": "JSON:API error document. The errors array contains one or more safe error objects; credential validation can return several field errors."
      },
...
      "ApplicationDocument": {
        "type": "object",
        "required": ["data"],
        "properties": {
          "data": {
            "type": "object",
            "required": ["type", "id", "attributes"],
            "properties": {
              "type": {
                "type": "string",
                "const": "applicant-applications"
              },
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "attributes": {
                "type": "object",
                "required": ["status"],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": ["draft", "submitted"],
                    "description": "Current lifecycle status. A submitted application also includes submittedAt and receipt."
                  },
                  "submittedAt": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Present only after successful submission."
                  },
                  "receipt": {
                    "type": "object",
                    "description": "Generic, non-sensitive next-step text; contains no review timeframe.",
                    "required": ["message", "nextStep"],
                    "properties": {
                      "message": {
                        "type": "string",
                        "const": "Your application was received"
                      },
                      "nextStep": {
                        "type": "string",
                        "const": "The review team will review your application."
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "description": "JSON:API response describing application lifecycle state and, after submission, its receipt."
      },
      "SubmissionRequestDocument": {
        "type": "object",
        "required": ["data"],
        "properties": {
          "data": {
            "type": "object",
            "required": ["type", "id", "attributes"],
            "properties": {
              "type": {
                "type": "string",
                "const": "applicant-applications"
              },
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "attributes": {
                "type": "object",
                "required": ["status", "confirmed", "expectedFormVersion"],
                "properties": {
                  "status": {
                    "type": "string",
                    "const": "submitted"
                  },
                  "confirmed": {
                    "type": "boolean",
                    "const": true
                  },
                  "expectedFormVersion": {
                    "type": "integer",
                    "minimum": 0,
                    "description": "Version returned by the latest form response. Prevents submission against stale form data."
                  }
                }
              }
            }
          }
        },
        "description": "JSON:API request document for explicit application submission."
      },
...
    },
    "responses": {
...
      "AuthenticationRequired": {
        "description": "No active Applicant session is available. Sign in and send the session cookie.",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDocument"
            },
            "examples": {
              "authenticationRequired": {
                "summary": "Session required",
                "value": {
                  "errors": [
                    {
                      "status": "401",
                      "code": "authentication-required",
                      "title": "Authentication required",
                      "detail": "Sign in to continue."
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "AccessDenied": {
        "description": "The request is not authorized for this resource or is missing a valid CSRF token. The response does not disclose protected resource details.",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDocument"
            },
            "examples": {
              "accessDenied": {
                "summary": "Access denied",
                "value": {
                  "errors": [
                    {
                      "status": "403",
                      "code": "access-denied",
                      "title": "Access denied",
                      "detail": "You do not have permission to access this resource."
                    }
                  ]
                }
              }
            }
          }
        }
      },
...
      "UnsupportedRequestMediaType": {
        "description": "The request Content-Type is not accepted for this operation. JSON:API request bodies must use application/vnd.api+json. For document-evidence uploads, use multipart/form-data; image validation failures return 422.",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDocument"
            },
            "examples": {
              "unsupportedMediaType": {
                "summary": "Unsupported request media type",
                "value": {
                  "errors": [
                    {
                      "status": "415",
                      "code": "unsupported-media-type",
                      "title": "Unsupported media type",
                      "detail": "Use application/vnd.api+json for JSON:API request bodies."
                    }
                  ]
                }
              },
              "mediaTypeParameters": {
                "summary": "Unsupported JSON:API media type parameters",
                "value": {
                  "errors": [
                    {
                      "status": "415",
                      "code": "unsupported-media-type",
                      "title": "Unsupported media type",
                      "detail": "Use application/vnd.api+json without media type parameters."
                    }
                  ]
                }
              }
            }
          }
        }
      },
...
      "NotAcceptable": {
        "description": "The requested response representation is not supported. Accept application/vnd.api+json.",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDocument"
            },
            "examples": {
              "notAcceptable": {
                "summary": "Unsupported response representation",
                "value": {
                  "errors": [
                    {
                      "status": "406",
                      "code": "not-acceptable",
                      "title": "Not acceptable",
                      "detail": "Accept application/vnd.api+json for JSON:API responses."
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "MalformedJsonApiDocument": {
        "description": "The JSON request body could not be parsed as a JSON:API document.",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDocument"
            },
            "examples": {
              "malformedDocument": {
                "summary": "Malformed JSON:API document",
                "value": {
                  "errors": [
                    {
                      "status": "400",
                      "code": "malformed-document",
                      "title": "Malformed JSON:API document",
                      "detail": "The request body is not a valid JSON:API document."
                    }
                  ]
                }
              }
            }
          }
        }
      },
...
      "ApplicationNotFound": {
        "description": "The application is absent or is not visible to the authenticated Applicant.",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDocument"
            },
            "examples": {
              "applicationNotFound": {
                "summary": "Application not found",
                "value": {
                  "errors": [
                    {
                      "status": "404",
                      "code": "application-not-found",
                      "title": "Application not found",
                      "detail": "The requested application is not available."
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "FormVersionConflict": {
        "description": "The form version changed or the application cannot be submitted in its current state. Reload the form before retrying.",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDocument"
            },
            "examples": {
              "formVersionConflict": {
                "summary": "Form version or lifecycle conflict",
                "value": {
                  "errors": [
                    {
                      "status": "409",
                      "code": "form-version-conflict",
                      "title": "Form changed",
                      "detail": "Reload the form and try again."
                    }
                  ]
                }
              }
            }
          }
        }
      },
...
      "SubmissionUnprocessable": {
        "description": "The submission request is invalid or the draft is incomplete. Incomplete applications include missing field names in the error detail.",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDocument"
            },
            "examples": {
              "invalidSubmission": {
                "summary": "Submission request rejected",
                "value": {
                  "errors": [
                    {
                      "status": "422",
                      "code": "unprocessable-form",
                      "title": "Form could not be saved",
                      "detail": "The submitted form could not be accepted."
                    }
                  ]
                }
              },
              "missingRequiredFields": {
                "summary": "Required fields missing",
                "value": {
                  "errors": [
                    {
                      "status": "422",
                      "code": "required-fields-missing",
                      "title": "Required fields are missing",
                      "detail": "Complete the required fields before submitting: name, dateOfBirth."
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "InternalError": {
        "description": "The request could not be completed. The response omits internal exception and persistence details.",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDocument"
            },
            "examples": {
              "internalError": {
                "summary": "Unexpected server error",
                "value": {
                  "errors": [
                    {
                      "status": "500",
                      "code": "internal-error",
                      "title": "Request failed",
                      "detail": "The request could not be completed."
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "The persistence operation failed. The response uses the safe fixed message shown below and omits database details.",
        "content": {
          "application/vnd.api+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDocument"
            },
            "examples": {
              "submissionUnavailable": {
                "summary": "Persistence unavailable",
                "value": {
                  "errors": [
                    {
                      "status": "503",
                      "code": "submission-unavailable",
                      "title": "Submission unavailable",
                      "detail": "Your application was not submitted. Please retry later."
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "parameters": {
      "ApplicationId": {
        "name": "applicationId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "UUID of the application to read or update. The application must be visible to the authenticated Applicant."
      }
    },
...
  }
}